Organize the core directory a bit
